refactor: use ternary operator instead of if statements in useGoalSummaryStore
